web/requirements.txt
channels~=3.0.3
channels_redis~=3.2.0
Django~=3.1.6
djangorestframework~=3.12.2
pyyaml~=5.4.1
